pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "regulationcoder"
version = "0.1.0"
description = "End-to-end pipeline converting regulation documents into compliance software"
readme = {text = "RegulationCoder - Converting regulations into compliance software", content-type = "text/plain"}
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "RegulationCoder Team" }]
dependencies = [
    "anthropic>=0.79.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "langdetect>=1.0.9",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "rich>=13.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10", "Pillow>=10.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
regulationcoder = "regulationcoder.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/regulationcoder"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
